[Acceptance] Split tests
